Narrow met location value checks for eggs s32->u16

+++ PKHeX.Core/Legality/Tables/Tables4.cs
@@ -99,7 +99,7 @@ public static partial class Legal
         500, // Park Ball
     });
 
-    internal static readonly HashSet<int> ValidMet_DP = new()
+    internal static readonly HashSet<ushort> ValidMet_DP = new()
     {
         // 063: Flower Paradise unreleased DP event
         // 079: Newmoon Island unreleased DP event
@@ -113,12 +113,12 @@ public static partial class Legal
         101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111,
     };
 
-    internal static readonly HashSet<int> ValidMet_Pt = new(ValidMet_DP)
+    internal static readonly HashSet<ushort> ValidMet_Pt = new(ValidMet_DP)
     {
         63, 79, 85, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125,
     };
 
-    internal static readonly HashSet<int> ValidMet_HGSS = new()
+    internal static readonly HashSet<ushort> ValidMet_HGSS = new()
     {
         080, 112, 113, 114, 115, 116,
         126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140,
@@ -129,9 +129,9 @@ public static partial class Legal
         221, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231, 232,      234,                               //233: Pokéwalker
     };
 
-    internal static readonly HashSet<int> ValidMet_4 = new(ValidMet_Pt.Concat(ValidMet_HGSS));
+    internal static readonly HashSet<ushort> ValidMet_4 = new(ValidMet_Pt.Concat(ValidMet_HGSS));
 
-    internal static readonly HashSet<int> GiftEggLocation4 = new()
+    internal static readonly HashSet<ushort> GiftEggLocation4 = new()
     {
         2009, 2010, 2011, 2013, 2014,
     };
